feat(Gamer) : Player와 Dealer사이의 중복 코드 제거
