+++ src/main/java/domain/gamer/Dealer.java
@@ -2,17 +2,13 @@ package domain.gamer;
 
 import domain.card.Card;
 import domain.card.CardDeck;
-import domain.card.Cards;
-import java.util.List;
 
-public class Dealer {
+public class Dealer extends Gamer {
 
     private static final int MIN_DEALER_SCORE = 16;
     private final CardDeck cardDeck;
-    private final Cards cards;
 
     public Dealer() {
-        this.cards = new Cards();
         this.cardDeck = new CardDeck();
         initDealerCard();
     }
@@ -31,13 +27,5 @@ public class Dealer {
 
     public boolean shouldDraw() {
         return cards.getMinGameScore() <= MIN_DEALER_SCORE;
-    }
-
-    public int getMaxGameScore() {
-        return cards.countMaxScore();
-    }
-
-    public List<String> getCardStatus() {
-        return cards.getCardsName();
     }
 }
