+++ src/main/java/domain/gamer/Gamer.java
@@ -1,0 +1,21 @@
+package domain.gamer;
+
+import domain.card.Cards;
+import java.util.List;
+
+public class Gamer {
+    
+    final Cards cards;
+
+    public Gamer() {
+        this.cards = new Cards();
+    }
+
+    public int getMaxGameScore() {
+        return cards.countMaxScore();
+    }
+
+    public List<String> getCardStatus() {
+        return cards.getCardsName();
+    }
+}
